Implement directive style quiz
There is a massive pain in attempting to inject data into the
repeated QuizController's on a page depending on what slide
that quiz is supposed to be, without having the slide and quiz
modules know about each other.

This commit alters the QuizController to be directly on the
directive and have attributes bound to the scope via the
bindToController feature for the controller to pick up.

The big negative is this isn't as cleanly unit testable and
indeed this commit breaks the tests. This will be fixed on the
next commit.